[MCSE-884][SecLib_RNG] Fix escaping local HashKeyBuffer address to a variable declared outside the function
+ Add missing cast in RNG_GetTrueRandomNumber function

+++ services/SecLib_RNG/RNG.c
@@ -294,7 +294,7 @@ int RNG_GetTrueRandomNumber(uint32_t *pRandomNo)
             break;
         }
 
-        rng_ret = RNG_Specific_GetRandomData((uint8_t *)pRandomNo, sizeof(uint32_t));
+        rng_ret = RNG_Specific_GetRandomData((uint8_t *)pRandomNo, (uint16_t)sizeof(uint32_t));
         if (rng_ret <= 0)
         {
             status = gRngInternalError_d;
